ColonR1/requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.43.0
certifi==2025.8.3
charset-normalizer==3.4.3
datasets==4.1.1
deepspeed==0.15.4
dill==0.4.0
einops==0.8.0
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
h5py==3.15.0
hf-transfer==0.1.4
hf-xet==1.1.10
hjson==3.1.0
huggingface-hub==0.35.0
idna==3.10
inquirerpy==0.3.4
jinja2==3.1.4
joblib==1.5.2
latex2sympy2-extended==1.10.2
liger-kernel==0.5.2
markdown-it-py==4.0.0
markupsafe==2.1.5
math-verify==0.8.0
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
networkx==3.3
ninja==1.13.0
numpy==2.1.2
packaging==23.0
pandas==2.3.2
peft==0.17.1
pfzy==0.3.4
pillow==11.0.0
prompt-toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic-core==2.33.2
pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
regex==2025.9.18
requests==2.32.5
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.1.1
sentencepiece==0.1.99
six==1.17.0
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.4
tqdm==4.67.1
transformers==4.51.0
trl==0.14.0
typing-extensions==4.12.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.20.1